SpringBoot_Funkos/src/test/java/org/example/springboot_funkos/categoria/mappers/CategoriaMapperTest.java
package org.example.springboot_funkos.categoria.mappers;

import org.example.springboot_funkos.rest.categoria.dto.CategoriaDto;
import org.example.springboot_funkos.rest.categoria.mappers.CategoriaMapper;
import org.example.springboot_funkos.rest.categoria.model.Categoria;
import org.junit.jupiter.api.Test;

import java.time.LocalDateTime;

import static org.junit.jupiter.api.Assertions.*;

class CategoriaMapperTest {

    private final CategoriaMapper mapper = new CategoriaMapper();

    @Test
    void toCategoria() {
        CategoriaDto categoriaDto = new CategoriaDto();
        categoriaDto.setNombre("DISNEY");
        categoriaDto.setActivado(true);

        var res = mapper.toCategoria(categoriaDto);

        assertAll(
                () -> assertEquals(categoriaDto.getNombre(), res.getNombre()),
                () -> assertEquals(categoriaDto.getActivado(), res.getActivado())
        );
    }

    @Test
    void toCategoriaUpdate() {
        CategoriaDto categoriaDto = new CategoriaDto();
        categoriaDto.setNombre("DISNEY");
        categoriaDto.setActivado(true);

        Categoria categoria = new Categoria(
                null,
                "DISNEY",
                LocalDateTime.now(),
                LocalDateTime.now(),
                categoriaDto.getActivado()
        );

        var res = mapper.toCategoriaUpdate(categoriaDto, categoria);

        assertAll(
                () -> assertNull(res.getId()),
                () -> assertEquals(categoriaDto.getId(), res.getId()),
                () -> assertEquals(categoriaDto.getNombre(), res.getNombre()),
                () -> assertEquals(categoriaDto.getActivado(), res.getActivado())
        );
    }
}